@@ -2,6 +2,7 @@ package com.unipet7.mcommerce.fragments;
 
 import android.annotation.SuppressLint;
 import android.app.Activity;
+import android.content.Context;
 import android.content.Intent;
 import android.graphics.Canvas;
 import android.graphics.Paint;
@@ -10,20 +11,18 @@ import android.graphics.drawable.Drawable;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.LayoutInflater;
-import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
+import android.view.Window;
 import android.widget.Button;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.activity.result.ActivityResultLauncher;
+import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AlertDialog;
-import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.content.res.AppCompatResources;
-import androidx.appcompat.widget.Toolbar;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.ItemTouchHelper;
@@ -33,24 +32,31 @@ import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.unipet7.mcommerce.R;
-import com.unipet7.mcommerce.activities.PaymentMethodDetails;
+import com.unipet7.mcommerce.activities.DetailCheckout;
 import com.unipet7.mcommerce.activities.VoucherActivity;
 import com.unipet7.mcommerce.adapters.CartAdapter;
 import com.unipet7.mcommerce.databinding.FragmentCartBinding;
 import com.unipet7.mcommerce.firebase.FireStoreClass;
 import com.unipet7.mcommerce.models.ProductCart;
+import com.unipet7.mcommerce.models.Voucher;
+import com.unipet7.mcommerce.utils.Constants;
 
 import java.util.ArrayList;
 import java.util.Map;
+import java.util.Objects;
 
-public class fragment_cart extends Fragment {
+public class CartList extends Fragment {
     private static final int REQUEST_CODE_VOUCHER = 1;
+
+    private boolean isVoucherUpdated = false;
     private String voucherCode;
     private double voucherNumb;
     private double voucherMaxDiscount;
     private double voucherMiniumValue;
 
     FragmentCartBinding binding;
+
+    Voucher voucher;
     ArrayList<ProductCart> productCarts;
     private double totalCartPrice = 0.0;
     CartAdapter adapter;
@@ -68,6 +74,9 @@ public class fragment_cart extends Fragment {
                 AlertDialog.Builder builder = new AlertDialog.Builder(getContext());
                 LayoutInflater inflater = LayoutInflater.from(getContext());
                 View dialogView = inflater.inflate(R.layout.dialogmessage, null);
+                Window window = builder.create().getWindow();
+                assert window != null;
+                window.setBackgroundDrawableResource(android.R.color.transparent);
                 builder.setView(dialogView);
 
                 TextView dialogMessage = dialogView.findViewById(R.id.tv_message_details_dialog);
@@ -78,27 +87,23 @@ public class fragment_cart extends Fragment {
                 btnConfirm.setText("Xóa");
 
                 final AlertDialog alertDialog = builder.create();
+                Objects.requireNonNull(alertDialog.getWindow()).setBackgroundDrawableResource(android.R.color.transparent);
 
-                btnCancel.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        alertDialog.dismiss();
-                        if (lastSwipedPosition != -1) {
-                            adapter.notifyItemChanged(lastSwipedPosition);
-                            lastSwipedPosition = -1;
-                        }
+
+                btnCancel.setOnClickListener(v -> {
+                    alertDialog.dismiss();
+                    if (lastSwipedPosition != -1) {
+                        adapter.notifyItemChanged(lastSwipedPosition);
+                        lastSwipedPosition = -1;
                     }
                 });
                 dialogMessage.setText("Bạn có muốn xóa sản phẩm khỏi giỏ hàng?");
                 dialogTitle.setText("Xóa sản phẩm");
 
-                btnConfirm.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        String cartItemId = String.valueOf(productCarts.get(position).getProductId());
-                        deleteItem(cartItemId, position);
-                        alertDialog.dismiss();
-                    }
+                btnConfirm.setOnClickListener(v -> {
+                    String cartItemId = String.valueOf(productCarts.get(position).getProductId());
+                    deleteItem(cartItemId, position);
+                    alertDialog.dismiss();
                 });
                 alertDialog.show();
                 lastSwipedPosition = position;
@@ -147,43 +152,74 @@ public class fragment_cart extends Fragment {
             FireStoreClass fireStoreClass = new FireStoreClass();
             fireStoreClass.deleteCartItem(productId);
             calculateTotalCartPrice();
-            CalculateVoucher();
+            CalculateVoucher(voucherCode, voucherNumb, voucherMaxDiscount, voucherMiniumValue);
         }
     };
 
 
     ItemTouchHelper itemTouchHelper;
 
+
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         binding = FragmentCartBinding.inflate(inflater, container, false);
-        setActionBar(binding.toolbarall);
+        loadData();
+        if (adapter != null) {
+            adapter.setOnQuantityChangeListener((position, quantity) -> {
+                ProductCart productCart = productCarts.get(position);
+                productCart.setNumOfProduct((int) quantity);
+                productCart.setTotalPrice(productCart.getProductPrice() * quantity);
+                adapter.updateCartItem(productCart);
+                adapter.notifyItemChanged(position);
+                calculateTotalCartPrice();
+            });
+        }
         itemTouchHelper = new ItemTouchHelper(simpleCallback);
         itemTouchHelper.attachToRecyclerView(binding.rcCart);
         binding.btnPaymentNow.setOnClickListener(v -> {
-            Intent intent = new Intent(getActivity(), PaymentMethodDetails.class);
+            Intent intent = new Intent(getActivity(), DetailCheckout.class);
+            Bundle bundle = new Bundle();
+            bundle.putDouble("totalCartPrice", totalCartPrice);
+            intent.putExtras(bundle);
             startActivity(intent);
         });
 
-        binding.btnCircleVoucher.setOnClickListener(v -> startActivityForResult(new Intent(getContext(), VoucherActivity.class), REQUEST_CODE_VOUCHER));
+        ActivityResultLauncher<Intent> voucherActivityResultLauncher = registerForActivityResult(
+                new ActivityResultContracts.StartActivityForResult(),
+                result -> {
+                    if (result.getResultCode() == Activity.RESULT_OK) {
+                        Intent data = result.getData();
+                        if (data != null) {
+                            voucher = (Voucher) data.getSerializableExtra(Constants.VOUCHER);
+                            assert voucher != null;
+                            this.voucherCode = voucher.getVoucher_code();
+                            this.voucherNumb = voucher.getVoucher_numb();
+                            this.voucherMaxDiscount = voucher.getVoucher_max_discount();
+                            this.voucherMiniumValue = voucher.getVoucher_minium_value();
+                            isVoucherUpdated = true;
+                            Log.i("CartList", "onActivityResult: " + voucherCode);
+                            CalculateVoucher(voucher.getVoucher_code(), voucher.getVoucher_numb(), voucher.getVoucher_max_discount(), voucher.getVoucher_minium_value());
+                        }
+                    }
+                });
+
+        binding.btnCircleVoucher.setOnClickListener(v -> voucherActivityResultLauncher.launch(new Intent(getContext(), VoucherActivity.class)));
         binding.btnVoucher.setOnClickListener(v -> {
             String voucherCode = binding.edtVoucher.getText().toString().trim();
             if (voucherCode.isEmpty()) {
-                startActivityForResult(new Intent(getContext(), VoucherActivity.class), REQUEST_CODE_VOUCHER);
+                voucherActivityResultLauncher.launch(new Intent(getContext(), VoucherActivity.class));
             } else {
                 checkVoucherExistence(voucherCode);
             }
         });
+
         FireStoreClass fireStoreClass = new FireStoreClass();
-        fireStoreClass.getCartItemsRealtime(getContext(), new CartAdapter.OnQuantityChangeListener() {
-            @Override
-            public void onQuantityChange(int position, double quantity) {
-                ProductCart productCart = productCarts.get(position);
-                productCart.setNumOfProduct((int) quantity);
-                productCart.setTotalPrice(productCart.getProductPrice() * quantity);
-                adapter.notifyItemChanged(position);
-                calculateTotalCartPrice();
-            }
+        fireStoreClass.getCartItemsRealtime(getContext(), (position, quantity) -> {
+            ProductCart productCart = productCarts.get(position);
+            productCart.setNumOfProduct((int) quantity);
+            productCart.setTotalPrice(productCart.getProductPrice() * quantity);
+            adapter.notifyItemChanged(position);
+            calculateTotalCartPrice();
         });
 
         return binding.getRoot();
@@ -217,7 +253,7 @@ public class fragment_cart extends Fragment {
                                 Toast.makeText(getContext(), "Dữ liệu voucher min không hợp lệ.", Toast.LENGTH_SHORT).show();
                                 return;
                             }
-                            CalculateVoucher();
+                            CalculateVoucher(voucherCode, voucherNumb, voucherMaxDiscount, voucherMiniumValue);
                         } else {
                             Toast.makeText(getContext(), "Dữ liệu voucher chung không hợp lệ.", Toast.LENGTH_SHORT).show();
                         }
@@ -234,7 +270,7 @@ public class fragment_cart extends Fragment {
 
     public void updateCartItem(ProductCart productCart) {
         FirebaseFirestore db = FirebaseFirestore.getInstance();
-        db.collection("cart")
+        db.collection(Constants.CART)
                 .whereEqualTo("productId", productCart.getProductId())
                 .whereEqualTo("userId", productCart.getUserId())
                 .get()
@@ -245,29 +281,6 @@ public class fragment_cart extends Fragment {
                     }
                 });
         calculateTotalCartPrice();
-        CalculateVoucher();
-    }
-
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        loadData();
-
-        if (adapter != null) {
-            adapter.setOnQuantityChangeListener(new CartAdapter.OnQuantityChangeListener() {
-                @Override
-                public void onQuantityChange(int position, double quantity) {
-                    ProductCart productCart = productCarts.get(position);
-                    productCart.setNumOfProduct((int) quantity);
-                    productCart.setTotalPrice(productCart.getProductPrice() * quantity);
-                    adapter.updateCartItem(productCart);
-                    adapter.notifyItemChanged(position);
-                    calculateTotalCartPrice();
-                    CalculateVoucher();
-                }
-            });
-        }
     }
 
 
@@ -275,8 +288,8 @@ public class fragment_cart extends Fragment {
         FirebaseFirestore db = FirebaseFirestore.getInstance();
         String currentUserId = new FireStoreClass().getCurrentUID();
         if (currentUserId != null && !currentUserId.isEmpty()) {
-            db.collection("cart")
-                    .whereEqualTo("userId", currentUserId)
+            db.collection(Constants.CART)
+                    .whereEqualTo(Constants.USER_ID, currentUserId)
                     .get()
                     .addOnSuccessListener(queryDocumentSnapshots -> {
                         productCarts = new ArrayList<>();
@@ -306,57 +319,37 @@ public class fragment_cart extends Fragment {
         for (ProductCart cartItem : productCarts) {
             totalCartPrice += cartItem.getTotalPrice();
         }
-        binding.txtPreNumb.setText(String.valueOf(Math.round(totalCartPrice) + " đ"));
-
-        // Check if a voucher is applied
-        String voucherCode = binding.edtVoucher.getText().toString().trim();
-        if (!voucherCode.isEmpty()) {
-        } else {
-            binding.txtTotalNumb.setText(String.valueOf(Math.round(totalCartPrice) + " đ"));}
+        String formatTotalCartPrice = String.format("%,.0f đ", totalCartPrice);
+        binding.txtPreNumb.setText(formatTotalCartPrice);
+        updateCartPriceDisplay(totalCartPrice, 0.0);
     }
 
-
-    public void setActionBar(@Nullable Toolbar toolbar) {
-        ((AppCompatActivity) requireActivity()).setSupportActionBar(toolbar);
-        ActionBar actionBar = ((AppCompatActivity) requireActivity()).getSupportActionBar();
-        if (actionBar != null) {
-            actionBar.setDisplayShowTitleEnabled(false);
+    private void CalculateVoucher(String voucherCode, double voucherNumb, double voucherMaxDiscount, double voucherMinimumValue) {
+        binding.edtVoucher.setText(voucherCode);
+        if (voucherCode.isEmpty()) {
+            updateCartPriceDisplay(totalCartPrice, 0.0);
+            return;
+        }
+        if (totalCartPrice >= voucherMinimumValue) {
+            double discountPrice = Math.min(totalCartPrice * voucherNumb / 100, voucherMaxDiscount);
+            double finalPrice = totalCartPrice - discountPrice;
+            updateCartPriceDisplay(finalPrice, discountPrice);
+        } else {
+            double valueNeed = voucherMinimumValue - totalCartPrice;
+            updateCartPriceDisplay(totalCartPrice, 0.0);
+            Toast.makeText(getContext(), "Bạn cần mua thêm " + String.format("%,.0f đ", valueNeed) + " để sử dụng voucher.", Toast.LENGTH_SHORT).show();
         }
     }
+
+    private void updateCartPriceDisplay(double finalPrice, double discountPrice) {
+        binding.txtDiscoutNumb.setText(String.format("%,.0f đ", discountPrice));
+        binding.txtTotalNumb.setText(String.format("%,.0f đ", finalPrice));
+    }
+
+
 
     @Override
-    public void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        if (requestCode == REQUEST_CODE_VOUCHER && resultCode == Activity.RESULT_OK && data != null) {
-            voucherCode = data.getStringExtra("voucher_code");
-            voucherNumb = data.getDoubleExtra("voucher_numb", 0.0);
-            voucherMaxDiscount = data.getDoubleExtra("voucher_max", 0.0);
-            voucherMiniumValue = data.getDoubleExtra("voucher_minium_value", 0.0);
-            CalculateVoucher();
-        }
-    }
-
-    private void CalculateVoucher() {
-        if (totalCartPrice >= voucherMiniumValue) {
-            double discountPrice = totalCartPrice * voucherNumb / 100;
-            if (discountPrice <= voucherMaxDiscount) {
-                binding.edtVoucher.setText(voucherCode);
-                double finalPrice = totalCartPrice - discountPrice;
-                binding.txtDiscoutNumb.setText(String.valueOf(Math.round(discountPrice)) + " đ");
-                binding.txtTotalNumb.setText(String.valueOf(Math.round(finalPrice)) + " đ");
-            } else {
-                binding.edtVoucher.setText(voucherCode);
-                binding.txtDiscoutNumb.setText(String.valueOf(Math.round(voucherMaxDiscount)) + " đ");
-                double finalPrice = totalCartPrice - voucherMaxDiscount;
-                binding.txtTotalNumb.setText(String.valueOf(Math.round(finalPrice)) + " đ");
-            }
-        }
-        else {
-            binding.edtVoucher.setText("Nhập mã giảm giá");
-            double valueNeed = voucherMiniumValue - totalCartPrice;
-            binding.txtDiscoutNumb.setText(String.valueOf(Math.round(0.0)) + " đ");
-            binding.txtTotalNumb.setText(String.valueOf(Math.round(totalCartPrice)) + " đ");
-            Toast.makeText(getContext(), "Bạn cần mua thêm " + String.valueOf(Math.round(valueNeed)) + " đ để sử dụng voucher.", Toast.LENGTH_SHORT).show();
-        }
+    public void onAttach(@NonNull Context context) {
+        super.onAttach(context);
     }
 }
