@@ -15,17 +15,16 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.RadioButton;
-import android.widget.RadioGroup;
 
 import com.unipet7.mcommerce.R;
-import com.unipet7.mcommerce.adapters.PaymentMethodListener;
+import com.unipet7.mcommerce.utils.DetailCheckoutHelper;
 import com.unipet7.mcommerce.databinding.FragmentPaymentMethodBinding;
 
 import org.checkerframework.checker.nullness.qual.NonNull;
 
 public class FragmentPaymentMethod extends Fragment {
     FragmentPaymentMethodBinding binding;
-    private PaymentMethodListener paymentMethodListener;
+    private DetailCheckoutHelper paymentMethodListener;
 
 
     @Override
@@ -43,8 +42,8 @@ public class FragmentPaymentMethod extends Fragment {
     @Override
     public void onAttach(@NonNull Context context) {
         super.onAttach(context);
-        if (context instanceof PaymentMethodListener) {
-            paymentMethodListener = (PaymentMethodListener) context;
+        if (context instanceof DetailCheckoutHelper) {
+            paymentMethodListener = (DetailCheckoutHelper) context;
         }
     }
 
@@ -57,28 +56,20 @@ public class FragmentPaymentMethod extends Fragment {
         savedCheckedRadioButton.setChecked(true);
 
         // Lắng nghe sự kiện thay đổi của RadioGroup
-        binding.paymentRadio.setOnCheckedChangeListener(new RadioGroup.OnCheckedChangeListener() {
-            @Override
-            public void onCheckedChanged(RadioGroup group, int checkedId) {
-                // Lưu giá trị RadioButton được chọn vào SharedPreferences
-                SharedPreferences.Editor editor = sharedPreferences.edit();
-                editor.putInt("checkedRadioButtonId", checkedId);
-                editor.apply();
-                // Gọi phương thức onPaymentMethodSelected của listener và truyền phương thức thanh toán được chọn
-                RadioButton checkedRadioButton = binding.paymentRadio.findViewById(checkedId);
-                String selectedPaymentMethod = checkedRadioButton.getText().toString();
-                if (paymentMethodListener != null) {
-                    paymentMethodListener.onPaymentMethodSelected(selectedPaymentMethod);
-                }
+        binding.paymentRadio.setOnCheckedChangeListener((group, checkedId) -> {
+            // Lưu giá trị RadioButton được chọn vào SharedPreferences
+            SharedPreferences.Editor editor = sharedPreferences.edit();
+            editor.putInt("checkedRadioButtonId", checkedId);
+            editor.apply();
+            // Gọi phương thức onPaymentMethodSelected của listener và truyền phương thức thanh toán được chọn
+            RadioButton checkedRadioButton = binding.paymentRadio.findViewById(checkedId);
+            String selectedPaymentMethod = checkedRadioButton.getText().toString();
+            if (paymentMethodListener != null) {
+                paymentMethodListener.onPaymentMethodSelected(selectedPaymentMethod);
             }
         });
 
-        binding.btnChoosePayment.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                requireActivity().getOnBackPressedDispatcher().onBackPressed();
-            }
-        });
+        binding.btnChoosePayment.setOnClickListener(v -> requireActivity().getOnBackPressedDispatcher().onBackPressed());
     }
 
     public void setActionBar(@Nullable Toolbar toolbar) {
@@ -90,10 +81,5 @@ public class FragmentPaymentMethod extends Fragment {
         actionBar.setDisplayShowTitleEnabled(false);
     }
     private void addEvents() {
-        binding.toolbar.setNavigationOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                requireActivity().getOnBackPressedDispatcher().onBackPressed();
-            }
-        });
+        binding.toolbar.setNavigationOnClickListener(v -> requireActivity().getOnBackPressedDispatcher().onBackPressed());
 }}
