@@ -1,11 +1,15 @@
 package com.unipet7.mcommerce.activities;
 
 import android.os.Bundle;
+import android.os.Handler;
 import android.util.Log;
+import android.widget.Toast;
+
+import androidx.activity.OnBackPressedCallback;
+import androidx.activity.OnBackPressedDispatcher;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentTransaction;
 import androidx.recyclerview.widget.RecyclerView;
 import androidx.viewpager2.widget.ViewPager2;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
@@ -14,13 +18,12 @@ import com.google.firebase.firestore.FirebaseFirestore;
 import com.unipet7.mcommerce.R;
 import com.unipet7.mcommerce.adapters.MainViewPager2Adapter;
 import com.unipet7.mcommerce.databinding.ActivityMainBinding;
+import com.unipet7.mcommerce.fragments.CartOverall;
 import com.unipet7.mcommerce.fragments.FragmentAllProduct;
-import com.unipet7.mcommerce.fragments.Fragment_Wishlist_Product;
+import com.unipet7.mcommerce.fragments.FavoriteOverall;
 import com.unipet7.mcommerce.fragments.Home;
 import com.unipet7.mcommerce.fragments.Profile;
-import com.unipet7.mcommerce.fragments.fragment_cart;
 import com.unipet7.mcommerce.utils.Constants;
-import com.unipet7.mcommerce.utils.LoadingDialog;
 import com.unipet7.mcommerce.utils.NonSwipeAbleViewPager;
 
 public class MainActivity extends AppCompatActivity {
@@ -29,14 +32,15 @@ public class MainActivity extends AppCompatActivity {
     private static final String TAG = "MainActivity";
     ActivityMainBinding binding;
 
-    private ViewPager2 mainViewPager2;
+    public ViewPager2 mainViewPager2;
 
-    LoadingDialog loadingDialog;
-    BottomNavigationView bottomNavigationView;
+    public BottomNavigationView bottomNavigationView;
     FloatingActionButton fabCart;
-    private FirebaseFirestore db;
 
     Fragment fragment = null;
+
+    private boolean doubleBackToExitPressedOnce = false;
+    private final Runnable doubleBackToExitRunnable = () -> doubleBackToExitPressedOnce = false;
 
 
     @Override
@@ -51,8 +55,37 @@ public class MainActivity extends AppCompatActivity {
         MainViewPager2Adapter adapter = new MainViewPager2Adapter(this);
         mainViewPager2.setAdapter(adapter);
         navigateFragment();
+
+        boolean isFromProductDetail = getIntent().getBooleanExtra(Constants.FROM_PRODUCT_DETAIL, false);
+        boolean orderSuccess = getIntent().getBooleanExtra(Constants.ODERSUCCESS, false);
+
+        if (isFromProductDetail || orderSuccess) {
+            mainViewPager2.setCurrentItem(2);
+        }
+
+        final OnBackPressedDispatcher dispatcher = getOnBackPressedDispatcher();
+        dispatcher.addCallback(this, new OnBackPressedCallback(true) {
+            @Override
+            public void handleOnBackPressed() {
+                if (doubleBackToExitPressedOnce) {
+                    setEnabled(false);
+                    dispatcher.onBackPressed();
+                    return;
+                }
+
+                doubleBackToExitPressedOnce = true;
+                Toast.makeText(MainActivity.this, "Nhấn lần nữa để thoát", Toast.LENGTH_SHORT).show();
+
+                new Handler().postDelayed(doubleBackToExitRunnable, 2000);
+            }
+        });
     }
 
+    @Override
+    protected void onDestroy() {
+        super.onDestroy();
+        new Handler().removeCallbacks(doubleBackToExitRunnable);
+    }
 
     private void mapping() {
         bottomNavigationView = binding.bottomNavigationView;
@@ -61,7 +94,7 @@ public class MainActivity extends AppCompatActivity {
         RecyclerView recyclerView = (RecyclerView) mainViewPager2.getChildAt(0);
         recyclerView.addOnItemTouchListener(new NonSwipeAbleViewPager());
         fabCart = binding.fabCart;
-        db = FirebaseFirestore.getInstance();
+        FirebaseFirestore db = FirebaseFirestore.getInstance();
     }
 
 
@@ -75,21 +108,21 @@ public class MainActivity extends AppCompatActivity {
                         fragment = new Home();
                         bottomNavigationView.getMenu().getItem(0).setChecked(true);
                         break;
-
                     case 1:
                         fragment = new FragmentAllProduct();
-                        // truyền bundle category allproduct cho fragment allproduct
+                        // truyền bundle category all product cho fragment allproduct
                         Bundle bundle = new Bundle();
                         bundle.putString(Constants.CATEGORY, Constants.ALLPRODUCT);
                         fragment.setArguments(bundle);
                         bottomNavigationView.getMenu().getItem(1).setChecked(true);
                         break;
                     case 2:
-                        fragment = new fragment_cart();
+                        fragment = new CartOverall();
+                        // set other menu item to false
                         bottomNavigationView.getMenu().getItem(2).setChecked(true);
                         break;
                     case 3:
-                        fragment = new Fragment_Wishlist_Product();
+                        fragment = new FavoriteOverall();
                         bottomNavigationView.getMenu().getItem(3).setChecked(true);
                         break;
                     case 4:
@@ -105,8 +138,6 @@ public class MainActivity extends AppCompatActivity {
             if (id == R.id.home_icon_bottom) {
                 removeAllFragments();
                 mainViewPager2.setCurrentItem(0);
-
-
             } else if (id == R.id.product_icon_bottom) {
                 mainViewPager2.setCurrentItem(1);
             } else if (id == R.id.fab_nav_shop) {
@@ -128,14 +159,5 @@ public class MainActivity extends AppCompatActivity {
         fragmentManager.popBackStackImmediate(null, FragmentManager.POP_BACK_STACK_INCLUSIVE);
         fragmentManager.executePendingTransactions();
     }
-    @Override
-    public void onBackPressed() {
-        super.onBackPressed();
-        BottomNavigationView bottomNavigationView = findViewById(R.id.bottomNavigationView);
-        bottomNavigationView.setSelectedItemId(R.id.home_icon_bottom);
-    }
-
-
-
 
 }
