@@ -1,10 +1,7 @@
 package com.unipet7.mcommerce.adapters;
 
-import android.app.Activity;
 import android.content.Intent;
 import android.graphics.Paint;
-import android.os.Bundle;
-import android.os.Handler;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -15,46 +12,60 @@ import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import androidx.activity.result.ActivityResultLauncher;
 import androidx.annotation.NonNull;
-import androidx.fragment.app.FragmentActivity;
-import androidx.fragment.app.FragmentManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
+import com.google.firebase.firestore.FirebaseFirestore;
 import com.unipet7.mcommerce.R;
 import com.unipet7.mcommerce.activities.DetailProduct;
+import com.unipet7.mcommerce.firebase.CheckFavoriteHelper;
 import com.unipet7.mcommerce.firebase.FireStoreClass;
 import com.unipet7.mcommerce.models.Product;
 import com.unipet7.mcommerce.utils.Constants;
+import com.unipet7.mcommerce.utils.OnRemoveAllFavHelper;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class ProductAdapter extends RecyclerView.Adapter<ProductAdapter.ViewHolder> implements FavProductInterface {
+public class ProductAdapter extends RecyclerView.Adapter<ProductAdapter.ViewHolder> {
 
     List<Product> productList;
 
     FireStoreClass fireStoreClass = new FireStoreClass();
+
+    private FirebaseFirestore firestore;
+
+    private CheckFavoriteHelper checkFavoriteHelper;
+
+    String userId = fireStoreClass.getCurrentUID();
     private List<Integer> favList = new ArrayList<>();
 
-    @Override
-    public void onDataLoaded(List<Integer> favList) {
-        this.favList = favList;
-        notifyDataSetChanged();
-    }
-    public ProductAdapter(List<Product> productList) {
+    private boolean loadFav;
+
+
+    public ProductAdapter(List<Product> productList, FireStoreClass fireStoreClass, boolean loadFav) {
         this.productList = productList;
-        fireStoreClass.getFavoriteList(this);
+        this.userId = fireStoreClass.getCurrentUID();
+        this.checkFavoriteHelper = new CheckFavoriteHelper();
+        this.loadFav = loadFav;
+        fireStoreClass.addFavoriteChangeListener(() -> notifyDataSetChanged());
     }
+
+    @Override
+    public void onDetachedFromRecyclerView(@NonNull RecyclerView recyclerView) {
+        super.onDetachedFromRecyclerView(recyclerView);
+        fireStoreClass.removeFavoriteChangeListener(()-> notifyDataSetChanged());
+    }
+
     public void setData(List<Product> newData) {
         productList = newData;
     }
 
+
     @NonNull
     @Override
-    public ProductAdapter.ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        // TODO: Inflate your layout and return your view holder
+    public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.item_layout, parent, false);
         return new ViewHolder(view);
     }
@@ -62,77 +73,79 @@ public class ProductAdapter extends RecyclerView.Adapter<ProductAdapter.ViewHold
     @Override
     public void onBindViewHolder(@NonNull ProductAdapter.ViewHolder holder, int position) {
         Product product = productList.get(position);
-        holder.productname.setText(product.getProductname());
+        String userId = fireStoreClass.getCurrentUID();
+        String productIdString = String.valueOf(product.getProductId());
+        checkFavoriteHelper.checkUserFavorite(userId, productIdString, isFavorite -> holder.favorite.setChecked(isFavorite));
+        int productId = product.getProductId();
+        holder.favorite.setChecked(product.isFavorite());
+        String productName = product.getProductname();
+        holder.productname.setText(productName);
         double roundRating = Math.round(product.getProductratenum() * 10) / 10.0;
         holder.productratenum.setText("4.5");
         holder.numOfRating.setText("  (130)");
+        double price = product.getProductprice();
+        double presaleprice = product.getPresaleprice();
+        String formattedPrice = String.format("%,.0f đ", price);
+        String formattedPreSalePrice = String.format("%,.0f đ", presaleprice);
+        holder.btnAddCart.setOnClickListener(v -> {
+            String productImage = product.getProductImageUrl();
+            double numOfProduct = 1.0;
+            Log.d("DetailProduct", "productName: " + productName);
+            Log.d("DetailProduct", "productPrice: " + price);
+            Log.d("DetailProduct", "numOfProduct: " + numOfProduct);
+            Log.d("DetailProduct", "productImageUrl: " + productImage);
+            Log.d("DetailProduct", "productId: " + productId);
+            Toast.makeText(v.getContext(), "Thêm vào giỏ hàng thành công", Toast.LENGTH_SHORT).show();
+            fireStoreClass.addToCart(productId, productName, price, numOfProduct, productImage);
+        });
+
         if (product.getSalepercent() > 0) {
             int salepercent = (int) product.getSalepercent();
             holder.salepercent.setText("-"+ salepercent + " %");
-            double percent = product.getSalepercent();
-            double price = product.getProductprice();
-            double saleprice = price - (price * percent / 100);
-
-            String formattedPrice = String.format("%,.0f đ", price);
-            String formattedSalePrice = String.format("%,.0f đ", saleprice);
-
-            holder.productprice.setText(formattedSalePrice);
+            holder.productprice.setText(formattedPrice);
             holder.presaleprice.setPaintFlags(holder.presaleprice.getPaintFlags() | Paint.STRIKE_THRU_TEXT_FLAG);
-            holder.presaleprice.setText(formattedPrice);
-            holder.btnAddCart.setOnClickListener(v -> {
-                String productName = product.getProductname();
-                double productPrice = product.getProductprice() - (product.getProductprice()*product.getSalepercent() / 100);
-                String productImage = product.getProductImageUrl();
-                double productId = product.getProductId();
-                double numOfProduct = 1.0;
-                String userId = fireStoreClass.getCurrentUID();
-                Log.d("DetailProduct", "productName: " + productName);
-                Log.d("DetailProduct", "productPrice: " + productPrice);
-                Log.d("DetailProduct", "numOfProduct: " + numOfProduct);
-                Log.d("DetailProduct", "productImageUrl: " + productImage);
-                Log.d("DetailProduct", "productId: " + productId);
-                Toast.makeText(v.getContext(), "Thêm vào giỏ hàng thành công", Toast.LENGTH_SHORT).show();
-                fireStoreClass.addToCart(userId, productId, productName, productPrice,numOfProduct ,productImage);
-            });
-
+            holder.presaleprice.setText(formattedPreSalePrice);
         }else {
             holder.salepercent.setVisibility(View.GONE);
             holder.salespercentbg.setVisibility(View.GONE);
             holder.salebanner.setVisibility(View.INVISIBLE);
-            holder.productprice.setText(String.format("%,.0f đ", product.getProductprice()));
+            holder.productprice.setText(formattedPrice);
             holder.presaleprice.setVisibility(View.GONE);
-            holder.btnAddCart.setOnClickListener(v -> {
-                String productName = product.getProductname();
-                double productPrice = product.getProductprice() - (product.getProductprice()*product.getSalepercent() / 100);
-                String productImage = product.getProductImageUrl();
-                double productId = product.getProductId();
-                double numOfProduct = 1.0;
-                String userId = fireStoreClass.getCurrentUID();
-                Log.d("DetailProduct", "productName: " + productName);
-                Log.d("DetailProduct", "productPrice: " + productPrice);
-                Log.d("DetailProduct", "numOfProduct: " + numOfProduct);
-                Log.d("DetailProduct", "productImageUrl: " + productImage);
-                Log.d("DetailProduct", "productId: " + productId);
-                Toast.makeText(v.getContext(), "Thêm vào giỏ hàng thành công", Toast.LENGTH_SHORT).show();
-                fireStoreClass.addToCart(userId, productId, productName, productPrice,numOfProduct ,productImage);
-            });
-
         }
         Glide.with(holder.itemView.getContext()).load(product.getProductImageUrl()).into(holder.imvThumb);
 
         holder.itemView.setOnClickListener(v -> {
-            int productId = product.getProductId();
             Intent intent = new Intent(v.getContext(), DetailProduct.class);
             intent.putExtra(Constants.PRODUCT_ID, productId);
             v.getContext().startActivity(intent);
         });
-        holder.favorite.setOnCheckedChangeListener((buttonView, isChecked) -> {
-            if (isChecked)
-                fireStoreClass.addFavorite(buttonView.getContext(),product.getProductId());
-            else {
-                fireStoreClass.removeFavorite(buttonView.getContext(),product.getProductId());
+        holder.favorite.setOnClickListener(v -> {
+            // check if favorite is checked, then make it unchecked and remove from favorite list
+            boolean isChecked = holder.favorite.isChecked();
+            if (!isChecked) {
+                holder.favorite.setChecked(false);
+                fireStoreClass.removeFavorite(v.getContext(), productId);
+            } else {
+                holder.favorite.setChecked(true);
+                fireStoreClass.addFavorite(v.getContext(), productId);
             }
         });
+        // if loadFav == true, then when user click on favorite, it will remove the product from favorite list and remove from the view
+        if (loadFav) {
+            holder.favorite.setOnClickListener(v -> {
+                // check if favorite is checked, then make it unchecked and remove from favorite list
+                    fireStoreClass.removeFavorite(v.getContext(), productId);
+                    productList.remove(position);
+                    notifyItemRemoved(position);
+                    notifyItemRangeChanged(position, productList.size());
+                    if (productList.isEmpty()) {
+                        if (v.getContext() instanceof OnRemoveAllFavHelper) {
+                            boolean loadEmptyFav = true;
+                            ((OnRemoveAllFavHelper) v.getContext()).onRemoveFav(loadEmptyFav);
+                        }
+                    }
+            });
+        }
     }
 
     @Override
@@ -162,9 +175,5 @@ public class ProductAdapter extends RecyclerView.Adapter<ProductAdapter.ViewHold
             btnAddCart = itemView.findViewById(R.id.btnAddCart);
             favorite = itemView.findViewById(R.id.chkFavouriteItem);
         }
-    }
-
-    public interface OnItemClickListener {
-        void onItemClick(Product product);
     }
 }
