@@ -1,7 +1,9 @@
 package com.unipet7.mcommerce.fragments;
 
+import android.content.Context;
 import android.os.Bundle;
 
+import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
@@ -12,24 +14,20 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.unipet7.mcommerce.adapters.ProductAdapter;
-import com.unipet7.mcommerce.databinding.FragmentWishlistProductBinding;
+import com.unipet7.mcommerce.databinding.FragmentFavoriteListBinding;
 import com.unipet7.mcommerce.firebase.FireStoreClass;
 import com.unipet7.mcommerce.models.Product;
+import com.unipet7.mcommerce.utils.Constants;
+import com.unipet7.mcommerce.utils.LoadingDialog;
+import com.unipet7.mcommerce.utils.OnRemoveAllFavHelper;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class Fragment_Wishlist_Product extends Fragment {
+public class FavoriteList extends Fragment {
 
-    // TODO: Rename parameter arguments, choose names that match
-    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-    private static final String ARG_PARAM1 = "param1";
-    private static final String ARG_PARAM2 = "param2";
+    FragmentFavoriteListBinding binding;
 
-    // TODO: Rename and change types of parameters
-    private String mParam1;
-    private String mParam2;
-    FragmentWishlistProductBinding binding;
     public ProductAdapter fvAdapter;
 
     ArrayList<Product> favoriteList = new ArrayList<>();
@@ -40,39 +38,36 @@ public class Fragment_Wishlist_Product extends Fragment {
 
     FireStoreClass fireStoreClass = new FireStoreClass();
 
-    public Fragment_Wishlist_Product() {
-        // Required empty public constructor
-    }
+    LoadingDialog loadingDialog = new LoadingDialog();
 
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
-        }
+    private OnRemoveAllFavHelper onRemoveFavHelper;
+
+    Bundle bundle;
+
+    private boolean loadFav = false;
+
+    public FavoriteList() {
+        // Required empty public constructor
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        binding= FragmentWishlistProductBinding.inflate(inflater, container, false);
-        fvRecyclerView = binding.favlist;
+        binding = FragmentFavoriteListBinding.inflate(inflater, container, false);
+        this.fvRecyclerView = binding.favlist;
+        this.bundle = getArguments();
+        if (bundle != null) {
+            loadFav = bundle.getBoolean(Constants.LOADFAV, false);
+        }
         fireStoreClass.getUserFavorites(this,favoriteList);
         return binding.getRoot();
     }
 
-    @Override
-    public void onResume() {
-        super.onResume();
-        // remove the previous data
-        favoriteList.clear();
-        fireStoreClass.getUserFavorites(this,favoriteList);
-    }
-
     public void loadFavoriteProducts(ArrayList<Product> favProducts) {
-        fvAdapter = new ProductAdapter(favProducts);
+        fvRecyclerView.setVisibility(View.VISIBLE);
+        loadingDialog.dissmis();
+        fvAdapter = new ProductAdapter(favProducts,fireStoreClass, true);
         Log.d("TAG", "loadFavoriteProducts: "+ favProducts.size());
         fvRecyclerView.setLayoutManager(new GridLayoutManager(getContext(),2));
         fvRecyclerView.setAdapter(fvAdapter);
