@@ -4,7 +4,7 @@ public class Constants {
     public static final String USERS = "users";
     public static final String PRODUCTS = "PRODUCTS";
 
-    public static final String USER_ID = "id";
+    public static final String USER_ID = "userId";
 
     public static final String IS_SIGNED_IN = "isSignedIn";
     public static final String GREETED = "greeted";
@@ -24,4 +24,19 @@ public class Constants {
     public static final String FAVPRODUCTID = "favProductId";
     public static final String CATEGORYID = "categoryId";
     public static final String SALESPERCENT = "salepercent";
+    public static final String CART = "cart";
+
+    public static final String FROM_PRODUCT_DETAIL = "fromProductDetail";
+    public static final String FAVORITEBY = "isFavoriteBy";
+    public static final String NULL = "null";
+    public static final String ADDRESSES = "addresses";
+    public static final String ADDRESS = "address";
+    public static final String ADDRESSID = "addressId";
+    public static final String IS_FROM_CHECKOUT = "isFromCheckout";
+    public static final int RESULT_CODE_ADDRESS = 102;
+    public static final String AWAIT = "Chờ xác nhận";
+    public static final String ORDERS = "orders";
+    public static final String ODERSUCCESS = "orderSuccess";
+    public static final String LOADFAV = "loadFav";
+    public static final String VOUCHER = "voucher";
 }
